Move loading-instructions into src/core
