+++ src/core/loading-instructions.js
